+++ btclib/network.py
@@ -10,9 +10,10 @@
 
 """Network constants and associated functions."""
 
+from typing import List
+
 from .curve import Curve
 from .curves import secp256k1
-
 
 _NETWORKS = ['mainnet', 'testnet', 'regtest']
 
@@ -92,30 +93,46 @@ _P2WSH_PRV_PREFIXES = [MAIN_Zprv, TEST_Vprv, TEST_Vprv]
 _P2WSH_PUB_PREFIXES = [MAIN_Zpub, TEST_Vpub, TEST_Vpub]
 
 
+_PRV_VERSIONS_MAIN = [MAIN_xprv, MAIN_yprv, MAIN_zprv, MAIN_Yprv, MAIN_Zprv]
+_PRV_VERSIONS_TEST = [TEST_tprv, TEST_uprv, TEST_vprv, TEST_Uprv, TEST_Vprv]
+_PUB_VERSIONS_MAIN = [MAIN_xpub, MAIN_ypub, MAIN_zpub, MAIN_Ypub, MAIN_Zpub]
+_PUB_VERSIONS_TEST = [TEST_tpub, TEST_upub, TEST_vpub, TEST_Upub, TEST_Vpub]
+
+_PRV_VERSIONS = [_PRV_VERSIONS_MAIN, _PRV_VERSIONS_TEST, _PRV_VERSIONS_TEST]
+_PUB_VERSIONS = [_PUB_VERSIONS_MAIN, _PUB_VERSIONS_TEST, _PUB_VERSIONS_TEST]
+
+# TODO remove _REPEATED_NETWORKS as it provides false match for regtest
 _REPEATED_NETWORKS = [
     'mainnet', 'mainnet', 'mainnet', 'mainnet', 'mainnet',
     'testnet', 'testnet', 'testnet', 'testnet', 'testnet',
     'regtest', 'regtest', 'regtest', 'regtest', 'regtest']
-
-_PRV_VERSIONS = [
-    MAIN_xprv, MAIN_yprv, MAIN_zprv, MAIN_Yprv, MAIN_Zprv,
-    TEST_tprv, TEST_uprv, TEST_vprv, TEST_Uprv, TEST_Vprv,
-    TEST_tprv, TEST_uprv, TEST_vprv, TEST_Uprv, TEST_Vprv]
-_PUB_VERSIONS = [
-    MAIN_xpub, MAIN_ypub, MAIN_zpub, MAIN_Ypub, MAIN_Zpub,
-    TEST_tpub, TEST_upub, TEST_vpub, TEST_Upub, TEST_Vpub,
-    TEST_tpub, TEST_upub, TEST_vpub, TEST_Upub, TEST_Vpub]
+_PRV_VERSIONS_ALL = _PRV_VERSIONS_MAIN + _PRV_VERSIONS_TEST + _PRV_VERSIONS_TEST
+_PUB_VERSIONS_ALL = _PUB_VERSIONS_MAIN + _PUB_VERSIONS_TEST + _PUB_VERSIONS_TEST
 
 
 def curve_from_network(network: str) -> Curve:
     network_index = _NETWORKS.index(network)
     return _CURVES[network_index]
 
-def curve_from_bip32version(version: bytes) -> Curve:
-    if version in _PRV_VERSIONS:
-        network_index = _PRV_VERSIONS.index(version)
-    elif version in _PUB_VERSIONS:
-        network_index = _PUB_VERSIONS.index(version)
-    else:
-        raise ValueError(f'unknown version ({version!r})')
+
+def _network_index_from_bip32version(bip32version: bytes) -> int:
+    if bip32version in _PRV_VERSIONS_ALL:
+        return _PRV_VERSIONS_ALL.index(bip32version)
+    if bip32version in _PUB_VERSIONS_ALL:
+        return _PUB_VERSIONS_ALL.index(bip32version)
+    raise ValueError(f'Unknown BIP32 version ({bip32version!r})')
+
+
+def curve_from_bip32version(bip32version: bytes) -> Curve:
+    network_index = _network_index_from_bip32version(bip32version)
     return _CURVES[network_index]
+
+
+def _pub_versions_from_network(network: str) -> List[bytes]:
+    network_index = _NETWORKS.index(network)
+    return _PUB_VERSIONS[network_index]
+
+
+def _prv_versions_from_network(network: str) -> List[bytes]:
+    network_index = _NETWORKS.index(network)
+    return _PRV_VERSIONS[network_index]